// src/sessions/interfaces/session.interfaces.ts
export interface IUserCacheData {
  id: number
  email: string
}

export interface ICacheData {
  user: IUserCacheData
  refreshToken: string
}

export interface ISessionData {
  userId: number
  email: string
}

export interface IUpdateUserSessionsListData {
  sessions: string[]
  cacheData: ICacheData
}